dmenu: blank lines on stdin create 'hidden item' behavior
To quote from the manpage:

	If there is a blank line in stdin, all items after it are not
	listed as possible completions when there is no text entered. They
	become visible only if they match some entered text.

This is not guarded by a feature switch because dmenu didn't previously
do anything very sensible if there were blank lines in the input, ie
such input was never really valid before. Ergo we can do whatever we
want with it.

(And I have already added enough flags.)